[samp-local] - IDE support
