+++ src/commands/local/runtime-support/dotnet/ide-support/visualstudio/launchSettings.json
@@ -1,0 +1,10 @@
+{
+  "profiles": {
+    "HelloWorld": {
+      "commandName": "Executable",
+      "executablePath": "$(SolutionDir).samp-out\\bin\\Debug\\net6.0\\dotnet.exe",
+      "workingDirectory": "$(SolutionDir).samp-out",
+      "remoteDebugEnabled": false
+    }
+  }
+}
